+++ app/src/main/AndroidManifest.xml
@@ -24,7 +24,7 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme2">
         <activity
-            android:name=".home.KingsCrossActivity"
+            android:name=".KingsCrossActivity"
             android:label="@string/app_name"
             android:screenOrientation="portrait">
             <intent-filter>
@@ -35,7 +35,7 @@
         </activity>
 
         <activity
-            android:name=".items.ItemDetailActivity"
+            android:name=".ItemDetailActivity"
             android:screenOrientation="portrait"
             android:theme="@android:style/Theme.Holo.Light"></activity>
