+++ app/src/main/res/layout/fragment_item_list.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.nabilhachicha.kc.items.itemlist.ItemsFragment">
+    tools:context="com.nabilhachicha.kc.view.itemlist.ItemsFragment">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/item_list"
