+++ app/src/main/res/layout/fragment_main.xml
@@ -29,7 +29,7 @@
             android:layout_width="match_parent"
             android:layout_height="?attr/actionBarSize" />
 
-        <com.nabilhachicha.kc.home.view.SlidingTabLayout
+        <com.nabilhachicha.kc.view.header.SlidingTabLayout
             android:id="@+id/sliding_tabs"
             android:layout_width="match_parent"
             android:layout_height="wrap_content" />
